Adds warning on buffer underrun on Windows
Windows can silently drop frames if data isn't read quickly enough.
This change adds a warning for this event.

As it turns out, Windows actually does this kind of a lot, even when
matching block sizes and nframes are used. It seems that it is
advisable to use a longer block size than nframes on Windows.