implements heuristic for silent sound cards on windows
on Windows, some soundcards indicate silence by returning zeros.
Others however indicate silence by returning no data at all. This
commit adds a heuristic output that estimates an appropriate length of
zeros when the sound card is not returning any data.

The length of the zeros is based on measuring the wall-clock time, and
might be off by up to four device periods when the sound card
starts/stops playing. This is not usually a problem, however, since
we're only missing silence anyway.