fixes empty soundcard list on macos
The code allocates some memory, has coreaudio write data to that
memory, then return the data. This is fine so long as the data is
trivial. However, if the data is a cdata itself, the memory it points
to will be deallocated once the function returns, thus invalidating
the cdata.

This use-after-free manifested when reading the number of channels in
a soundcard, reporting them as 0, and discarding them. Thus no
soundcard could ever be found.

Unless, that is, the garbage collector had a slow day. In that case,
the channels were read before the data was collected, and everything
seemed fine. That's how things used to be in the past, so the bug
didn't surface most of the time.