fixes regression on channel property
For some reason, the mChannelsPerFrame in the
AudioStreamBasicDescription needs to be written at startup, but is
always reported as 0 on reading. This changed at some point in the
past, probably in some macOS update. So now we just store our own copy
of the number of channels. Such are the joys of macOS development.